requirements.txt
discord.py>=1.5.1
config~=0.5.0.post0